src/api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
claude-code-sdk>=0.0.23
python-multipart>=0.0.6